refactor(reusable-screens): PR changes
