+++ app/screens/reusable/Acknowledgement/AcknowledgementTitle/AcknowledgementTitle.types.ts
