+++ app/screens/reusable/Acknowledgement/AcknowledgementTitle/index.ts
@@ -1,2 +1,2 @@
 export { AcknowledgementTitleScreen } from './AcknowledgementTitle';
-export type { IAcknowledgementTitleRoute } from './AcknowledgementTitle.props';
+export type { IAcknowledgementTitleRoute } from './AcknowledgementTitle.types';
