+++ app/screens/reusable/Acknowledgement/AcknowledgementTitleSubtitle/AcknowledgementTitleSubtitle.types.ts
