+++ app/screens/reusable/Acknowledgement/AcknowledgementTitleSubtitle/index.ts
@@ -1,2 +1,2 @@
 export type { AcknowledgementTitleSubtitleType } from './AcknowledgementTitleSubtitle.data';
-export type { IAcknowledgementTitleSubtitleRoute } from './AcknowledgementTitleSubtitle.props';
+export type { IAcknowledgementTitleSubtitleRoute } from './AcknowledgementTitleSubtitle.types';
