+++ app/screens/reusable/Dialog/Dialog.types.ts
