+++ app/screens/reusable/EmojiSelection/index.ts
@@ -1,2 +1,2 @@
 export { EmojiSelectionScreen } from './EmojiSelection';
-export type { IEmojiSelectionRoute } from './EmojiSelection.props';
+export type { IEmojiSelectionRoute } from './EmojiSelection.types';
