+++ app/screens/reusable/VerticalSwipe/VerticalSwipe.types.ts
